requirements.txt
PyYAML==3.10
Twisted==12.0.0
boto==2.3.0
clint==0.3.1
distribute==0.6.24
gevent==0.13.7
gevent-zeromq==0.2.2
greenlet==0.3.4
pyasn1==0.1.3
pycrypto==2.5
pyzmq-static==2.1.7.1
zope.interface==3.8.0
